+++ src/main/java/io/github/astrapi69/string/package-info.java
@@ -1,6 +1,6 @@
 /**
  * This package contains helper classes for String objects <br>
- * 
+ *
  * @author Asterios Raptis
  */
 package io.github.astrapi69.string;